refactor: split relay server responsibilities
