+++ rust/alera-cli/src/terminal_host/server.rs
@@ -39,7 +39,6 @@ use crate::terminal_host::orchestration::message_delivery::{
 use crate::terminal_host::orchestration::message_formatter::format_messages_for_injection;
 use crate::terminal_host::orchestration::message_waiters::MessageWaiterRegistry;
 use crate::terminal_host::protocol::{event, TerminalHostConfig};
-use crate::terminal_host::relay_runtime;
 use crate::terminal_host::session::{PtyWriteCompletion, Session};
 use crate::terminal_host::sleep_detector::SleepDetector;
 
@@ -120,6 +119,7 @@ mod host_status;
 mod lifecycle;
 mod managed_workspace_requests;
 mod mobile_gateway_surface;
+mod mobile_hello_requests;
 mod mobile_terminal_requests;
 #[cfg(test)]
 mod mobile_terminal_viewport_tests;
@@ -144,6 +144,7 @@ mod prompt_image_store;
 mod pty_event_forwarder;
 mod pty_events;
 mod push_delivery;
+mod remote_relay;
 mod request_payloads;
 mod requests;
 mod resource_requests;
@@ -264,61 +265,6 @@ enum MobileGatewayReplacement {
 }
 
 impl ServerActor {
-    pub(super) async fn restart_remote_relay(&mut self) {
-        self.stop_remote_relay().await;
-        let settings = match self.runtime_store.mobile_access_settings().await {
-            Ok(settings) => settings,
-            Err(error) => {
-                tracing::warn!("could not read remote mobile access settings: {error}");
-                return;
-            }
-        };
-        if !settings.remote_access_enabled {
-            return;
-        }
-        match self.account_push.service.local_account().await {
-            Ok(Some(_)) => {}
-            Ok(None) => return,
-            Err(error) => {
-                tracing::warn!("could not read the local account for remote relay: {error}");
-                return;
-            }
-        }
-        let (task, stop) = relay_runtime::spawn(
-            self.account_push.service.clone(),
-            self.account_push.service.runtime_id().to_owned(),
-            self.inbox.clone(),
-            self.next_client_id.clone(),
-        );
-        self.account_push.relay_task = Some(task);
-        self.account_push.relay_stop = Some(stop);
-        self.cancel_shutdown_timer();
-        self.broadcast_authenticated(event("mobileRelayChanged", json!({ "enabled": true })));
-    }
-
-    pub(super) async fn stop_remote_relay(&mut self) {
-        if let Some(stop) = self.account_push.relay_stop.take() {
-            let _ = stop.send(());
-        }
-        if let Some(task) = self.account_push.relay_task.take() {
-            task.abort();
-            let _ = task.await;
-        }
-        self.dispose_relay_clients().await;
-        self.broadcast_authenticated(event("mobileRelayChanged", json!({ "enabled": false })));
-    }
-
-    async fn dispose_relay_clients(&mut self) {
-        let client_ids = self
-            .clients
-            .iter()
-            .filter_map(|(id, client)| client.relay_client_id.as_ref().map(|_| *id))
-            .collect::<Vec<_>>();
-        for client_id in client_ids {
-            self.dispose_client(client_id).await;
-        }
-    }
-
     pub(super) async fn restart_mobile_gateway(&mut self) -> HostResult<()> {
         let settings = self
             .runtime_store
